Make source map location information more readable.
This changes "file:1:2" to "line 1, column 2 of file" and ":1:2" to
"line 1, column 2" in [SourceFile.getLocationMessage]. The more
machine-readable colon-separated format doesn't make sense for a
message intended for humans.

Since [Location.formatString] seems more likely to be consumed by
machines, its format was left as-is.

This also prepares version 0.9.1 for release.

This was previously submitted as r36603. That caused buildbot errors and was
rolled back by r36615. This CL fixes the test errors in r36603.

R=sigmund@google.com

Review URL: https://codereview.chromium.org//304603003

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart/pkg/source_maps@36717 260f80e4-7a28-3924-810f-c04153c831b5