Add SpanException and SpanFormatException classes to source_maps.
It's often useful to associate source range information with an
exception in a way that will get printed along with that exception but
can be pulled out and manipulated directly. It's especially useful if
there's a common superclass/interface for doing so that can be shared
across packages. This adds that superclass.

R=sigmund@google.com

Review URL: https://codereview.chromium.org//348493004

git-svn-id: https://dart.googlecode.com/svn/branches/bleeding_edge/dart/pkg/source_maps@37509 260f80e4-7a28-3924-810f-c04153c831b5